Update salary page to new design
This commit updates the salary page to the new design. It structures the text differently, adds graphs for pension, bonus and salary, and refactors some surrounding code.

Note: There is still a lot of work to do on this page. The structure of the html is quite strange, as the setup in sanity is not optimal. In sanity right now, we have to duplicate all benefits, as each benefit is related to one city. We want to create a shared benefit type that can be shared with all cities. This would also make the structure in the html better.

I also want to note that there are some todos left in this code. Right now things work, but some updates are neccesary to reduce maintainance, and make the code easier to work with.