Make handbook its own section
Previously, the handbook was integrated into the compensationCalculator section, or the generocity section. This ment that the same data was places multiple places, and it was impossible to have just a handbook component without the calculator.

The handbook was removed from the calculator in a previous commit, so that change is not visible here. i have just moved the handbook component out of the file. To clearify, it was never used in this state, just sitting in the calculator file. Now it has its own section file.

I have also used this new section in the generocity section.