Add split section as section option
This commit adds split section as a section option. This was previously only a component used in the sharedStudio, but i have now decided to add it to the normal studio, with some modifications.

Currently, the split section only supports calculator and handbook. This can be extended in the future if wanted.

This split section was added as we needed it at the bottom of the /jobs page for the calculator and handbook placement