Refactor sanity salaries
This commit refactors salaries in sanity to not be related to locations. It has been decided that there is no reason for each location to have their own salaries. Now, there are common salaries for all locations.

It also implements a more robust way of fetching the correct data, as we now sort all options, and fetch the first one. This makes it so that the order of how the salaries are added in sanity does not matter, as they are sorted in the query to ensure that the latest one are always fetched